Catch ConnectExceptions when shutting down the server
It is possible for the server thread to end before its next call to
`accept`. In such a scenario, a java.net.ConnectException is thrown.
This exception is now being caught. Additionally, the shutdown function
is checking if the thread is still alive before attempting to stop it
from listening.